requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
chromadb
faiss_cpu
langchain_google_genai
openpyxl
fpdf
youtube_transcript_api
pathlib
pytube
opencv-python